适配 ROLL 点插件
