+++ src/plugins/roll/plugins/rand/data_source.py
@@ -1,5 +1,4 @@
-""" 掷骰子
-"""
+""" 掷骰子 """
 import re
 from random import randint
